pyproject.toml
[project]
name = "tinygp"
description = "The tiniest of Gaussian Process libraries"
authors = [{ name = "Dan Foreman-Mackey", email = "foreman.mackey@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
dynamic = ["version"]
dependencies = ["jax", "jaxlib", "equinox"]

[project.optional-dependencies]
test = ["pytest"]
comparison = ["george", "celerite"]
docs = [
    "sphinx-book-theme",
    "myst-nb",
    "matplotlib",
    "statsmodels",
    "numpyro",
    "arviz",
    "flax",
    "optax",
    "jaxopt",
    "ipython",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/tinygp/tinygp_version.py"

[tool.black]
target-version = ["py39"]
line-length = 88

[tool.ruff]
target-version = "py39"
line-length = 88
select = ["F", "I", "E", "W", "YTT", "B", "Q", "PLE", "PLR", "PLW", "UP"]
ignore = [
    "E741",    # Allow ambiguous variable names
    "PLR0911", # Allow many return statements
    "PLR0913", # Allow many arguments to functions
    "PLR0915", # Allow many statements
    "PLR2004", # Allow magic numbers in comparisons
]
exclude = []

[tool.ruff.isort]
known-first-party = ["tinygp"]
combine-as-imports = true

[tool.towncrier]
name = "tinygp"
package = "tinygp"
package_dir = "src"
filename = "docs/news.rst"
directory = "news/"
issue_format = "`#{issue} <https://github.com/dfm/tinygp/issues/{issue}>`_"
underlines = "-~~"
